Refactor: move constants to specific files
