+++ src/utils/ignored-function-names.ts
@@ -1,0 +1,7 @@
+// skip the reserved names: https://developers.google.com/apps-script/guides/html/reference/run
+export default [
+  'withFailureHandler',
+  'withLogger',
+  'withSuccessHandler',
+  'withUserObject',
+];
